added a random ai agent

+++ bmp/BMP.h
@@ -129,7 +129,7 @@ struct BMP {
             bmp_info_header.bit_count = 32;
             bmp_info_header.compression = 3;
             row_stride = width * 4;
-            data.resize(row_stride * height);
+            data.resize((size_t)row_stride * height);
             file_header.file_size = file_header.offset_data + data.size();
         }
         else {
@@ -139,7 +139,7 @@ struct BMP {
             bmp_info_header.bit_count = 24;
             bmp_info_header.compression = 0;
             row_stride = width * 3;
-            data.resize(row_stride * height);
+            data.resize((size_t)row_stride * height);
 
             uint32_t new_stride = make_stride_aligned(4);
             file_header.file_size = file_header.offset_data + static_cast<uint32_t>(data.size()) + bmp_info_header.height * (new_stride - row_stride);
